backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
google-generativeai
passlib[bcrypt]
python-jose
